Rename `constraint.py` -> `advanced_constraint.py`.
